Split main reconciliation logic and dns provider
Initial version was completely tailored to work with kubernetes ingresses
and how CERN Openstack managed landb-aliases work.

In the future we plan to support also services and load balancers and
LanDB official API. Also allowing to work with Oracle Cloud Infrastructure
and other external DNSs.

To start shifting towards that direction this commit separates the controller,
plan and provider logic so that the controller creates a plan and propagates
it to a generic provider. Specific provider implementeation will propagate to
the corresponding infra.

Signed-off-by: Guillermo Facundo Colunga <guillermo.facundo.colunga@cern.ch>