feat: Add Dockerfile and CI for automated builds
Introduces a multi-stage Dockerfile to create a minimal,
multi-architecture container for the application. The build
process uses `scratch` as the base image for minimal size.

The GitLab CI pipeline has been updated with a new `build`
and `tagging` stages that automatically build and tag
the Docker image in the GitLab registry.

The CI is implemented to push an image as follows:

- On non-master branch commit publish: `image:<commit_hash>`
- A `-rc` tag is added when a commit is added to default
branch: `image:<commit_hash>-rc`
- On tag -> `image:<tag>`

Signed-off-by: Luis Pigueiras <luis.pigueiras@cern.ch>