ci: split lint into multiple stages per language
This commit splits the linting stage into multiple parallel jobs.
Each job will lint its own language, for example at the moment
the repo is using YAML, Go and markdown files. Also a bunch of
gp files. So this commits adds a linting stage for general,
for go, for yaml and for makdown. Also added a speel checker
for all the markdown docs.

Signed-off-by: Guillermo Facundo Colunga <guillermo.facundo.colunga@cern.ch>